requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.2
pandas==2.2.2
numpy==2.0.1
websocket-client==1.8.0
requests==2.32.3
